+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Added Discord to the links section in settings. displaying these links in a visible area, such as a footer, so users can easily access them.
 - Added a button so you can retry sending an email to a user. by @nayskutzu
 - Added the you accept our terms of service checkbox to the login page. by @nayskutzu
+- Added a custom badge system to roles so you can set a custom badge for each role. by @nayskutzu
 
 ### Improved